move some shared code to separate module, update pre-commit cfg
