+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
       - id: check-useless-excludes
       - id: check-hooks-apply
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
+    rev: v4.4.0
     hooks:
       - id: check-added-large-files
       - id: check-case-conflict
@@ -37,13 +37,13 @@ repos:
         language_version: python3
 
   - repo: https://github.com/pycqa/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
       - id: isort
         args: ["--settings-path=pyproject.toml"]
 
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.961
+    rev: v1.4.1
     hooks:
     - id: mypy
       additional_dependencies:
@@ -59,16 +59,16 @@ repos:
       files: src/ymltoxml/ymltoxml.py
 
   - repo: "https://github.com/asottile/blacken-docs"
-    rev: "v1.12.1"
+    rev: "1.15.0"
     hooks:
       - id: "blacken-docs"
         name: "Format docs (blacken-docs)"
         args: ["-l", "64"]
         additional_dependencies:
-          - "black==21.9b0"
+          - "black==22.1.0"
 
   - repo: https://github.com/PyCQA/doc8
-    rev: 0.11.2
+    rev: v1.1.1
     hooks:
       - id: doc8
         args:
@@ -76,7 +76,7 @@ repos:
           - '--ignore=D001'
 
   - repo: https://github.com/pre-commit/pygrep-hooks
-    rev: v1.9.0
+    rev: v1.10.0
     hooks:
       - id: rst-backticks
         # exclude: ChangeLog\.rst$
@@ -84,7 +84,7 @@ repos:
       - id: rst-inline-touching-normal
 
   - repo: https://github.com/myint/autoflake
-    rev: v1.4
+    rev: v2.2.0
     hooks:
       - id: autoflake
         files: src/ymltoxml/ymltoxml.py
@@ -94,15 +94,15 @@ repos:
           - --remove-duplicate-keys
           - --remove-unused-variables
 
-  - repo: https://gitlab.com/pycqa/flake8
-    rev: 3.9.2
+  - repo: https://github.com/pycqa/flake8
+    rev: 6.0.0
     hooks:
       - id: flake8
         files: src/ymltoxml/ymltoxml.py
         additional_dependencies: ["flake8-bugbear"]
 
   - repo: https://github.com/PyCQA/bandit
-    rev: 1.7.4
+    rev: 1.7.5
     hooks:
       - id: bandit
         args: ["-ll", "-q"]
@@ -112,19 +112,3 @@ repos:
 #    rev: v6.2.1
 #    hooks:
 #      - id: beautysh
-
-ci:
-    autofix_commit_msg: |
-        [pre-commit.ci] auto fixes from pre-commit.com hooks
-
-        for more information, see https://pre-commit.ci
-    autofix_prs: false
-    autoupdate_commit_msg: '[pre-commit.ci] pre-commit autoupdate'
-    autoupdate_schedule: weekly
-    skip: []
-    submodules: false
-
-# re-running a pull request: you can trigger a re-run on a pull request by
-#    commenting pre-commit.ci run (must appear on a line by itself).
-# skipping push runs: skip a run by putting [skip ci], [ci skip],
-#    [skip pre-commit.ci], or [pre-commit.ci skip] in the commit message.
